# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления бронированиями",
        "title": "Booker Microservice API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/bookingsType/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить детальную информацию о типе бронирования",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingsType"
                ],
                "summary": "Получить тип бронирования по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID типа бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_get_booking_type_by_id.GetBookingTypeResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновить тип бронирования",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingsType"
                ],
                "summary": "Обновить тип бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID типа бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные типа бронирования",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_update_booking_type.UpdateBookingTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ShlykovPavel_booker_microservice_models_booking_type_create_booking_type.ResponseId"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удалить тип бронирования по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookingsType"
                ],
                "summary": "Удалить тип бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID типа бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ShlykovPavel_booker_microservice_models_booking_type_create_booking_type.ResponseId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "github_com_ShlykovPavel_booker_microservice_models_booking_type_get_booking_type_by_id.GetBookingTypeResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ShlykovPavel_booker_microservice_models_booking_type_update_booking_type.UpdateBookingTypeRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Add \"Bearer\" before token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}